requirements.txt
alembic==1.0.10
aniso8601==6.0.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
colorama==0.4.1
coverage==4.5.3
cryptography==2.6.1
dj-database-url==0.5.0
Flask==1.0.3
Flask-Cors==3.0.7
flask-marshmallow==0.10.1
Flask-Migrate==2.5.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
gunicorn==19.9.0
importlib-metadata==0.17
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
more-itertools==7.0.0
passlib==1.7.1
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.5.0
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
